MetaDoubleParamがdouble型になってない問題を修正
MetaDoubleParamはちゃんとdouble型を返すようにした

+++ Metasia.Core/Coordinate/CoordPoint.cs
@@ -15,7 +15,7 @@ namespace Metasia.Core.Coordinate
         /// <summary>
         /// 保持する値
         /// </summary>
-        public float Value = 0f;
+        public double Value = 0f;
 
         /// <summary>
         /// ポイント間の移動方法
