+++ Metasia.Core/Coordinate/PointLogic/StraightLineMove.cs
@@ -7,7 +7,7 @@ namespace Metasia.Core.Coordinate.PointLogic
     /// </summary>
     public class StraightLineMove : IPointLogic
     {
-        public float GetBetweenPoint(float StartValue, float EndValue, float NowFrame, float StartFrame, float EndFrame)
+        public double GetBetweenPoint(double StartValue, double EndValue, double NowFrame, double StartFrame, double EndFrame)
         {
             if(StartValue == EndValue) return StartValue;
             return StartValue + (EndValue - StartValue) * (NowFrame - StartFrame) / (EndFrame - StartFrame);
